pyproject.toml
[project]
name = "engliah-teacher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-allauth>=65.13.1",
    "django>=6.0",
    "django-environ>=0.12.0",
    "google-generativeai>=0.8.5",
    "markdown>=3.10",
    "django-pwa>=2.0.1",
    "django-dynamic-preferences>=1.17.0",
]
